requirements.txt
Flask>=2.3.0
androguard>=3.4.0
joblib>=1.3.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
requests>=2.31.0
loguru>=0.7.0
cryptography>=41.0.0
frida-tools>=12.0.0
frida>=16.0.0
threatingestor>=1.4.0
beautifulsoup4>=4.4.1
ipaddress>=1.0.18
iocextract>=1.12.1
PyYAML>=3.13
jsonpath_rw>=1.4.0
sgmllib3k>=1.0.0
statsd>=4.0.1
pyshorteners>=1.0.1
twitter>=2.0a2
httpretty>=1.1.4
responses>=0.25.8
mock>=5.2.0
boto3>=1.40.26
Nose>=1.3.7
sphinx>=8.2.3
Pygments>=2.19.2
collective.checkdocs>=0.2
coverage>=7.10.6
hug>=2.6.1
pymisp>=2.5.17
greenstalk==1.0.1
watchdog>=6.0.0
feedparser>=5.2.1
threatkb>=1.0.0a2
pymysql>=1.1.2
pandas>=2.1.0 
seaborn>=0.13.2
matplotlib>=3.10.6
imbalanced-learn>=0.14.0
tensorflow>=2.20.0
python-dotenv>=1.0.0